# config/swagger.yaml
openapi: 3.0.0
info:
  title: Project Dashboard API
  version: 1.0.0
  description: API documentation for the project dashboard backend.
servers:
  - url: http://localhost/api
    description: Local development server
...
paths:
...
  /report-links/{appId}:
    get:
      tags: [ReportLinks]
      summary: Get all report links
      parameters:
        - in: path
          name: appId
          required: true
          schema:
            type: string
          description: Application ID
      responses:
        '200':
          description: List of report links
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    nameFa:
                      type: string
                    nameEn:
                      type: string
                    url:
                      type: string
                    ownerId:
                      type: string
                    reportCategory:
                      type: string
                    reportSubCategory:
                      type: string
        '500':
          description: Server error

    post:
      tags: [ReportLinks]
      summary: Add a new report link
      parameters:
        - in: path
          name: appId
          required: true
          schema:
            type: string
          description: Application ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nameFa
                - nameEn
                - url
                - ownerId
                - reportCategory
                - reportSubCategory
              properties:
                nameFa:
                  type: string
                nameEn:
                  type: string
                url:
                  type: string
                ownerId:
                  type: string
                reportCategory:
                  type: string
                reportSubCategory:
                  type: string
      responses:
        '201':
          description: Report link added successfully
        '400':
          description: Missing required fields
        '500':
          description: Server error
  /report-links/{appId}/{linkId}:
    put:
      tags: [ReportLinks]
      summary: Update a report link
      parameters:
        - in: path
          name: appId
          required: true
          schema:
            type: string
          description: Application ID
        - in: path
          name: linkId
          required: true
          schema:
            type: string
          description: Report link ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nameFa
                - nameEn
                - url
                - ownerId
                - reportCategory
                - reportSubCategory
              properties:
                nameFa:
                  type: string
                nameEn:
                  type: string
                url:
                  type: string
                ownerId:
                  type: string
                reportCategory:
                  type: string
                reportSubCategory:
                  type: string
      responses:
        '200':
          description: Report link updated successfully
        '400':
          description: Missing required fields
        '404':
          description: Report link not found
        '500':
          description: Server error

    delete:
      tags: [ReportLinks]
      summary: Delete a report link
      parameters:
        - in: path
          name: appId
          required: true
          schema:
            type: string
          description: Application ID
        - in: path
          name: linkId
          required: true
          schema:
            type: string
          description: Report link ID
      responses:
        '200':
          description: Report link deleted successfully
        '404':
          description: Report link not found
        '500':
          description: Server error